разрешение конфликта слияния с game-with-host

+++ backend/modules/game_session/services.py
@@ -1,7 +1,8 @@
 from typing import List
 
 from backend.modules.game.repos import game_repo
-from backend.modules.game_session.dtos import GameStateDTO, GameSessionDescriptionDTO
+from backend.modules.game_session.dtos import GameStateDTO, GameSessionDescriptionDTO, CurrentQuestionAnswerDTO, \
+    HostGameStateDTO
 from backend.modules.game_session.entities import GameSession
 from backend.modules.game_session.events import GameSessionCreatedEvent, GameSessionDeletedEvent
 from backend.modules.game_session.exceptions import AlreadyPlaying, AlreadyCreated, GameSessionNotFound
@@ -16,18 +17,19 @@ class GameSessionService:
 
     def get_game_state(self, username: str) -> GameStateDTO:
         user = self.user_repo.get(username)
-
-        if not user.is_playing:
+        if user.is_playing:
+            game_session = self.repo.get(user.game_session_id)
+            return GameStateDTO(game_session)
+        elif user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+            return HostGameStateDTO(game_session)
+        else:
             raise GameSessionNotFound()
-
-        game_session = self.repo.get(user.game_session_id)
-
-        return GameStateDTO(game_session)
 
     def create(self, game_session_data, username: str) -> GameStateDTO:
         user = self.user_repo.get(username)
 
-        if user.is_playing:
+        if user.is_playing or user.is_hosting:
             raise AlreadyPlaying
 
         if self.repo.is_exists(user):
@@ -36,6 +38,7 @@ class GameSessionService:
         game = self.game_repo.get(game_session_data['game_name'])
 
         game_session = GameSession(creator=user,
+                                   host=user if game_session_data['is_host'] else None,
                                    game=game,
                                    max_players=game_session_data['max_players'])
 
@@ -64,28 +67,34 @@ class GameSessionService:
     def join(self, username: str, join_data):
         user = self.user_repo.get(username)
 
+        # TODO брать id сессии из creator.game_session_id
         game_session = self.repo.get_by_creator(join_data['creator'])
 
-        if not user.is_playing:
+        if not (user.is_playing or user.is_hosting):
             game_session.join(user)
 
             game_session = self.repo.save(game_session)
-        elif user.game_session_id != game_session.id:
+        elif game_session.id not in (user.game_session_id, user.hosted_game_session_id):
             raise AlreadyPlaying
 
         return GameStateDTO(game_session)
 
-    def leave(self, username: str):
+    def leave(self, username: str):  # TODO сообщать игрокам о выходе ведущего
         user = self.user_repo.get(username)
 
-        if not user.is_playing:
+        if user.is_playing:
+            game_session = self.repo.get(user.game_session_id)
+        elif user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+        else:
             raise GameSessionNotFound()
 
-        game_session = self.repo.get(user.game_session_id)
+        if not user.is_hosting:
+            game_session.leave(user)
 
-        game_session.leave(user)
-
-        if game_session.is_all_players_left():
+        # TODO отдельное уведомление о выходе ведущего
+        # TODO не удалять сессию при выходе ведущего
+        if user.is_hosting or not game_session.is_hosted and game_session.is_all_players_left():
             game_session.add_event(GameSessionDeletedEvent(game_session))
 
             print('gs deleted')
@@ -93,6 +102,17 @@ class GameSessionService:
             self.repo.delete(game_session)
         else:
             self.repo.save(game_session)
+
+    def start(self, username: str):
+        user = self.user_repo.get(username)
+
+        if user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+            game_session.start_game()
+
+            self.repo.save(game_session)
+        else:
+            raise GameSessionNotFound()
 
     def choose_question(self, username: str, question_data):
         user = self.user_repo.get(username)
@@ -109,6 +129,19 @@ class GameSessionService:
 
         self.repo.save(game_session)
 
+    def allow_answers(self, username: str):
+        user = self.user_repo.get(username)
+
+        if user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+            game_session.allow_answers()
+
+            self.repo.save(game_session)
+
+            return CurrentQuestionAnswerDTO(game_session)
+        else:
+            raise GameSessionNotFound()
+
     def answer_timeout(self, game_session_id: int):
         game_session = self.repo.get(game_session_id)
 
@@ -123,20 +156,39 @@ class GameSessionService:
 
         game_session.final_round_timeout()
 
-        game_session.add_event(GameSessionDeletedEvent(game_session))
+        if game_session.is_hosted:
+            self.repo.save(game_session)
+        else:
+            game_session.add_event(GameSessionDeletedEvent(game_session))
 
-        self.repo.delete(game_session)
-
-        print('game ended')
+            self.repo.delete(game_session)
 
     def submit_answer(self, username: str, answer_data):
         user = self.user_repo.get(username)
-
-        if not user.is_playing:
-            raise GameSessionNotFound()
-
         game_session = self.repo.get(user.game_session_id)
 
-        game_session.submit_answer(user, answer_data['answer'])
+        game_session.submit_answer(user, answer_data['answer'] if answer_data else None)
 
         self.repo.save(game_session)
+
+    def confirm_answer(self, username: str):
+        user = self.user_repo.get(username)
+
+        if user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+            game_session.confirm_answer()
+
+            self.repo.save(game_session)
+        else:
+            raise GameSessionNotFound()
+
+    def reject_answer(self, username: str):
+        user = self.user_repo.get(username)
+
+        if user.is_hosting:
+            game_session = self.repo.get(user.hosted_game_session_id)
+            game_session.reject_answer()
+
+            self.repo.save(game_session)
+        else:
+            raise GameSessionNotFound()
